build Docker base image

+++ git2jar/src/main/java/git2jar/Git2jarApp.java
@@ -59,7 +59,9 @@ public class Git2jarApp {
             .build()
             .boot();
         System.out.println("==== web mode ====");
-        FileService.deleteDir(Config.config.getJobsWorkDir());
+        if (Config.config.getJobsWorkDir().isDirectory()) {
+        	FileService.deleteDir(Config.config.getJobsWorkDir());
+        }
     }
     
     private static class WebRoutes extends RouteDefinitions {
